# openapi/financial.yaml
openapi: 3.0.3
info:
  title: Financial Service API
  version: 1.0.0
  description: API for financial domain operations including accounts, clients, invoices, transactions, and attachments
servers:
  - url: http://localhost:3001
    description: Local development server
security:
  - BearerAuth: []
paths:
  /api/financial/accounts:
    get:
      summary: List bank accounts
      operationId: listAccounts
      tags:
        - Accounts
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountListResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/financial/transactions:
    get:
      summary: List transactions
      operationId: listTransactions
      tags:
        - Transactions
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: accountId
          in: query
          schema:
            type: string
            description: Filter by account ID
        - name: type
          in: query
          schema:
            type: string
            enum: [debit, credit, transfer]
            description: Filter by transaction type
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, failed, cancelled]
            description: Filter by transaction status
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
            description: Start date for date range filter
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
            description: End date for date range filter
        - name: minAmount
          in: query
          schema:
            type: number
            description: Minimum transaction amount
        - name: maxAmount
          in: query
          schema:
            type: number
            description: Maximum transaction amount
        - name: search
          in: query
          schema:
            type: string
            description: Search in description and counterparty name
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/financial/categorize/auto:
    post:
      summary: Auto-categorize transactions
      operationId: autoCategorizeTransactions
      tags:
        - Transactions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transactionIds:
                  type: array
                  items:
                    type: string
                    format: uuid
              required:
                - transactionIds
      responses:
        '200':
          description: Auto-categorization completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  categorizedCount:
                    type: integer
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/financial/setup-bbva:
    post:
      summary: Setup BBVA connection
      operationId: setupBbvaConnection
      tags:
        - Setup
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Setup initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupResponse'
        '400':
          description: Setup failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/financial/complete-setup:
    post:
      summary: Complete setup process
      operationId: completeSetup
      tags:
        - Setup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                requisitionId:
                  type: string
              required:
                - requisitionId
      responses:
        '200':
          description: Setup completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupResult'
        '400':
          description: Invalid requisition ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/financial/sync:
    post:
      summary: Manual sync
      operationId: manualSync
      tags:
        - Sync
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResult'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/financial/dashboard/overview:
    get:
      summary: Get dashboard overview
      operationId: getDashboardOverview
      tags:
        - Dashboard
      security:
        - BearerAuth: []
      parameters:
        - name: currency
          in: query
          schema:
            type: string
            default: EUR
          description: Currency for financial data
      responses:
        '200':
          description: Dashboard overview data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardOverviewResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/financial/dashboard/quick-stats:
    get:
      summary: Get quick financial statistics
      operationId: getQuickStats
      tags:
        - Dashboard
      security:
        - BearerAuth: []
      parameters:
        - name: currency
          in: query
          schema:
            type: string
            default: EUR
          description: Currency for financial data
        - name: period
          in: query
          schema:
            type: string
          description: Time period for statistics
      responses:
        '200':
          description: Quick financial statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuickStatsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/financial/categories:
    get:
      summary: Get financial categories
      operationId: getCategories
      tags:
        - Dashboard
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, transfer]
          description: Filter by category type
      responses:
        '200':
          description: List of financial categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/financial/sync-status:
    get:
      summary: Get synchronization status
      operationId: getSyncStatus
      tags:
        - Dashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sync status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - message
        - code
      properties:
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code (BAD_REQUEST, NOT_FOUND, etc.)
...
    SetupResponse:
      type: object
      properties:
        success:
          type: boolean
        requisitionId:
          type: string
        authUrl:
          type: string
        status:
          type: string
    SetupResult:
      type: object
      properties:
        success:
          type: boolean
        accountsFound:
          type: integer
        message:
          type: string
    SyncResult:
      type: object
      properties:
        success:
          type: boolean
        synchronized:
          type: integer
        failed:
          type: integer
        lastSync:
          type: string
          format: date-time
        message:
          type: string
    PaginationInfo:
      type: object
      required:
        - page
        - limit
        - total
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
...
    AccountListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationInfo'
        - type: object
          required:
            - accounts
          properties:
            accounts:
              type: array
              items:
                $ref: '#/components/schemas/Account'
...
    TransactionListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationInfo'
        - type: object
          required:
            - transactions
          properties:
            transactions:
              type: array
              items:
                $ref: '#/components/schemas/Transaction'
            stats:
              type: object
              properties:
                totalIncome:
                  type: number
                totalExpenses:
                  type: number
...
    DashboardOverviewResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/DashboardMetrics'
        message:
          type: string
...
    CategoriesResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            required:
              - id
              - name
              - type
              - color
            properties:
              id:
                type: string
              name:
                type: string
              type:
                type: string
                enum: [income, expense, transfer]
              color:
                type: string
              icon:
                type: string
                nullable: true
        message:
          type: string
...